+++ dwds/lib/src/readers/proxy_server_asset_reader.dart
@@ -7,10 +7,10 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:logging/logging.dart';
+import 'package:dwds/src/utilities/shared.dart' show LogWriter;
 import 'package:shelf/shelf.dart';
 import 'package:shelf_proxy/shelf_proxy.dart';
 
-import '../../dwds.dart' show LogWriter;
 import 'asset_reader.dart';
 
 /// A reader for resources provided by a proxy server.